# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * INNDOS API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type SignupInputRole =
  (typeof SignupInputRole)[keyof typeof SignupInputRole];

export const SignupInputRole = {
  owner: "owner",
  tenant: "tenant",
  admin: "admin",
  host: "host",
  guest: "guest",
} as const;

export interface SignupInput {
  name: string;
  email: string;
  password: string;
  role?: SignupInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  owner: "owner",
  tenant: "tenant",
  admin: "admin",
  host: "host",
  guest: "guest",
} as const;

export type UserProfileStatus =
  (typeof UserProfileStatus)[keyof typeof UserProfileStatus];

export const UserProfileStatus = {
  active: "active",
  pending: "pending",
  suspended: "suspended",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
  status: UserProfileStatus;
  joinDate: string;
  avatar?: string | null;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType];

export const PropertyType = {
  rent: "rent",
  sale: "sale",
  bnb: "bnb",
  hotel: "hotel",
  hostel: "hostel",
} as const;

export type PropertyDetails = { [key: string]: unknown };

export type PropertyPropertyStatus =
  (typeof PropertyPropertyStatus)[keyof typeof PropertyPropertyStatus];

export const PropertyPropertyStatus = {
  pending: "pending",
  approved: "approved",
  flagged: "flagged",
  sold: "sold",
} as const;

export interface Property {
  id: string;
  ownerId: string;
  title: string;
  type: PropertyType;
  price: number;
  address: string;
  beds: number;
  baths: number;
  sqft: number;
  guests?: number | null;
  image: string;
  images: string[];
  videos: string[];
  details: PropertyDetails;
  isVerified: boolean;
  tags: string[];
  lat?: string | null;
  lng?: string | null;
  createdAt?: string;
  description?: string | null;
  subtype?: string | null;
  hourlyRate?: number | null;
  priceUnit?: string | null;
  /** @minimum 1 */
  totalUnits: number;
  propertyStatus?: PropertyPropertyStatus;
  activeBookingsCount?: number;
  ownerPhone?: string | null;
  ownerEmail?: string | null;
  ownerAvatar?: string | null;
  ownerBusinessName?: string | null;
  savedAt?: string | null;
  ownerName?: string | null;
}

export type CreatePropertyInputType =
  (typeof CreatePropertyInputType)[keyof typeof CreatePropertyInputType];

export const CreatePropertyInputType = {
  rent: "rent",
  sale: "sale",
  bnb: "bnb",
  hotel: "hotel",
  hostel: "hostel",
} as const;

export type CreatePropertyInputDetails = { [key: string]: unknown };

export interface CreatePropertyInput {
  title: string;
  type: CreatePropertyInputType;
  price: number;
  address: string;
  beds?: number;
  baths?: number;
  sqft?: number;
  guests?: number;
  image?: string;
  images?: string[];
  videos?: string[];
  details?: CreatePropertyInputDetails;
  tags?: string[];
  description?: string | null;
  lat?: string;
  lng?: string;
  subtype?: string;
  hourlyRate?: number;
  priceUnit?: string;
  /** @minimum 1 */
  totalUnits?: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export type ListingDraftData = { [key: string]: unknown };

export interface ListingDraft {
  id: string;
  userId: string;
  data: ListingDraftData;
  createdAt: string;
  updatedAt: string;
}

export type SaveListingDraftInputData = { [key: string]: unknown };

export interface SaveListingDraftInput {
  data: SaveListingDraftInputData;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: string;
  propertyId: string;
  userId: string;
  status: BookingStatus;
  startDate: string;
  endDate: string;
  totalPrice: number;
  createdAt?: string;
  propertyTitle?: string | null;
  propertyAddress?: string | null;
  propertyImage?: string | null;
  propertyImages?: string[] | null;
  propertyType?: string | null;
}

export interface FavoriteInput {
  propertyId: string;
}

export interface FavoriteStatus {
  isFavorited: boolean;
  propertyId: string;
}

export type BookedRangeStatus =
  (typeof BookedRangeStatus)[keyof typeof BookedRangeStatus];

export const BookedRangeStatus = {
  pending: "pending",
  confirmed: "confirmed",
  blocked: "blocked",
} as const;

export interface BookedRange {
  startDate: string;
  endDate: string;
  status: BookedRangeStatus;
}

export interface CreateBookingInput {
  propertyId: string;
  startDate: string;
  endDate: string;
  totalPrice: number;
}

export type ListPropertiesParams = {
  type?: ListPropertiesType;
  search?: string;
  ownerId?: string;
  minLat?: number;
  maxLat?: number;
  minLng?: number;
  maxLng?: number;
};

export type ListPropertiesType =
  (typeof ListPropertiesType)[keyof typeof ListPropertiesType];

export const ListPropertiesType = {
  rent: "rent",
  sale: "sale",
  bnb: "bnb",
  hotel: "hotel",
  hostel: "hostel",
} as const;