# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: INNDOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List current user's bookings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    patch:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        startDate:
          type: string
        endDate:
          type: string
        totalPrice:
          type: integer
        createdAt:
          type: string
        propertyTitle:
          type: string
          nullable: true
        propertyAddress:
          type: string
          nullable: true
        propertyImage:
          type: string
          nullable: true
        propertyImages:
          type: array
          items:
            type: string
          nullable: true
        propertyType:
          type: string
          nullable: true
      required:
        - id
        - propertyId
        - userId
        - status
        - startDate
        - endDate
        - totalPrice
...
    CreateBookingInput:
      type: object
      properties:
        propertyId:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalPrice:
          type: integer
      required:
        - propertyId
        - startDate
        - endDate
        - totalPrice